### 2019-11-01 #1861 @whedon
[PRE REVIEW]: univariateML: An R package for maximum likelihood estimation of univariate densities
**Submitting author:** @JonasMoss (<a href="http://orcid.org/0000-0002-6876-6964">Jonas Moss</a>)
**Repository:** <a href="https://github.com/JonasMoss/univariateML" target ="_blank">https://github.com/JonasMoss/univariateML</a>
**Version:** v1.0.0
