### 2019-11-02 #1861 @arfon
:wave: @JonasMoss - thanks for your submission to JOSS. Could you take a look at [this list of people](https://bit.ly/joss-reviewers) who have volunteered to review for JOSS and see if there are any good candidates?
