### 2019-11-02 #1861 @vbaliga
Hi @arfon and @JonasMoss - I am willing but I should mention that the next week or so is looking pretty busy. I don't see info in the JOSS guidelines about how long a window reviewers are given to return feedback. Would completing my review closer to mid- or end-November be OK? If you'd like to process this in a more timely manner, I won't be offended if you'd rather ask someone else.
