### 2019-11-02 #1863 @whedon
[REVIEW]: univariateML: An R package for maximum likelihood estimation of univariate densities
**Submitting author:** @JonasMoss (<a href="http://orcid.org/0000-0002-6876-6964">Jonas Moss</a>)
**Repository:** <a href="https://github.com/JonasMoss/univariateML" target ="_blank">https://github.com/JonasMoss/univariateML</a>
**Version:** v1.0.0

I see there's been some discussion about the License formatting but that @arfon has instructed to keep things as they are. So I will just check this off.

Yes, although I note that using either `devools::check()` or `devtools::install_github("JonasMoss/univariateML", build_vignettes = TRUE)` fails when attempting to create vignettes. This is likely because of how `copula` is needed but not installed & loaded by the "Copula Modeling" vignette. I suggest revising how `copula` is handled within this vignette just so that users who want local vignettes don't have a hard time.

Language such as "quick, easy, and free of bugs" is used in a few places, but I don't see any specifics on how the custom-made optimizers perform. I am familiar enough with many of them that I can attest to their veracity in fitting parameters (and they are well-covered by the package's `testthat`), but speed or efficiency haven't been documented. I do agree that `univariateML` provides a lot of convenience. But if language like "quick" is going to be used, it should be backed up with evidence.

I'm happy enough with the writing to check off the box here, but please note that other items in this review do relate to this topic.

This is generally covered well, but I'll make a few suggestions to enhance clarity and/or make life easier for people who are encountering the package for the first time.

Please link to your documentation site more directly within the README of the github repository. I realize there's a Documentation header inside the readme, but I'd also link to the main site (https://univariateml.netlify.com/index.html) at the top of the github repo (i.e. above the Topics) as well as within the Overview section.

Within the documentation site, please re-order the Articles dropdown menu so that the vignettes are listed in an order that's friendly to new users (i.e. put the Overview vignette first). This can be done by further editing the `_pkgdown.yml` in the root directory.

Within the Distributions vignette, I'd make it a little more clear that the Estimator column refers to functions within your `univariateML` package. I think this can be done via renaming the column and reordering the column names to: Name, univariateML Function, Parameters, Support, Package.

In the Copula vignette, the last line of the first section notes that the package can be used to "speed up task 1" (fitting marginal distributions), but the specifics of how speedy this is (e.g. with run times) or how it compares to approaches via other packages isn't shown. Again, I think that if speed or efficiency are invoked, there should be evidence shown. Or the language could just be revised.

I'm happy with what I see here! Very nicely thought out. I did not catch any errors.

I like the update you made following the other reviewer's comments. Looks good!

I'm happy enough with the language to check this off, but I'll also note that the first sentence of the paper could be tweaked to be a bit more descriptive. I like the language used in both the README and the Overview vignette -- it's nice to have a lead statement that gives me a full sense of what the package does in a punchy way. But this is a minor comment and not one that I'm looking to have enforced.

I'd recommend expanding this further. Right now, the paper mentions how this package compares to `stats4::mle()` but does not really go into enough detail about the specific benefits of using `univariateML`'s functions.

Within the paper, can you give at least one example of a custom-made optimizer within `univariateML` and show how this optimizer's performance fares against what is built into base R or another package? How have other functions let us down?

I suspect that the ability to fit each distribution is already available through other R packages (and this seems to be somewhat confirmed by the table in the Distributions vignette). If there are any distributions that are not covered by other packages, please make this explicit in the paper by specifying the name of the `univariateML` estimator function.

I also think there is merit to incorporating more of a "one stop shop" argument in here. It's convenient to have all of these functions within the same package, especially as they feed nicely into the "Data analysis aids" functions, which I'd argue are are the most important parts of the package. So I recommend that the language of the paper be further adjusted to tell the reader why it's good to use this package instead of relying on what is already out there.
