+++ README.Rmd
@@ -60,10 +60,10 @@ lines(mlweibull(egypt$age))
 
 ## Documentation
 For an overview of the package and its features see the 
-[overview vignette](/vignettes/overview.html). For a list of implemented 
-densities see the start of the [details vignette](/vignettes/distributions.html). 
+[overview vignette](https://univariateml.netlify.com/articles/overview.html). For a list of implemented 
+densities see the start of the [details vignette](https://univariateml.netlify.com/articles/distributions.html). 
 For an illustration of how this package can make an otherwise long and laborious process
-much simpler, see the [copula vignette](/vignettes/copula.html).
+much simpler, see the [copula vignette](https://univariateml.netlify.com/articles/copula.html).
 
 ## How to Contribute or Get Help
 If you encounter a bug, have a feature request or need some help, open a [Github issue](https://github.com/JonasMoss/univariateML/issues). 
