+++ README.md
@@ -57,11 +57,12 @@ lines(mlweibull(egypt$age))
 ## Documentation
 
 For an overview of the package and its features see the [overview
-vignette](/vignettes/overview.html). For a list of implemented densities
-see the start of the [details vignette](/vignettes/distributions.html).
+vignette](https://univariateml.netlify.com/articles/overview.html). For
+a list of implemented densities see the start of the [details
+vignette](https://univariateml.netlify.com/articles/distributions.html).
 For an illustration of how this package can make an otherwise long and
 laborious process much simpler, see the [copula
-vignette](/vignettes/copula.html).
+vignette](https://univariateml.netlify.com/articles/copula.html).
 
 ## How to Contribute or Get Help
 
