+++ R/mlnorm.R
@@ -23,6 +23,7 @@
 
 mlnorm = function(x, na.rm = FALSE) {
   if(na.rm) x = x[!is.na(x)] else assertthat::assert_that(!anyNA(x))
+  ml_input_checker(x)
   n = length(x)
 
   mu = mean(x)
