+++ R/mllnorm.R
@@ -24,6 +24,7 @@
 
 mllnorm = function(x, na.rm = FALSE) {
   if(na.rm) x = x[!is.na(x)] else assertthat::assert_that(!anyNA(x))
+  ml_input_checker(x)
   assertthat::assert_that(min(x) > 0)
   y = log(x)
   n = length(x)
