+++ R/mlbeta.R
@@ -29,7 +29,7 @@ mlbeta = function(x, na.rm = FALSE, start = NULL,
                   type = c("none", "gradient", "hessian")) {
 
   if(na.rm) x = x[!is.na(x)] else assertthat::assert_that(!anyNA(x))
-
+  ml_input_checker(x)
   assertthat::assert_that(min(x) > 0)
   assertthat::assert_that(max(x) < 1)
 
