+++ R/mllgamma.R
@@ -30,6 +30,7 @@ mllgamma = function(x, na.rm = FALSE, rel.tol = .Machine$double.eps^0.25,
                    iterlim = 100) {
 
   if(na.rm) x = x[!is.na(x)] else assertthat::assert_that(!anyNA(x))
+  ml_input_checker(x)
   assertthat::assert_that(min(x) > 1)
 
   y = log(x)
