+++ R/mlbetapr.R
@@ -32,6 +32,7 @@ mlbetapr = function(x, na.rm = FALSE, start = NULL,
                     type = c("none", "gradient", "hessian")) {
 
   if(na.rm) x = x[!is.na(x)] else assertthat::assert_that(!anyNA(x))
+  ml_input_checker(x)
   assertthat::assert_that(min(x) > 0)
 
   val1 = mean(log(x))
