+++ R/mlexp.R
@@ -22,6 +22,7 @@
 
 mlexp = function(x, na.rm = FALSE) {
   if(na.rm) x = x[!is.na(x)] else assertthat::assert_that(!anyNA(x))
+  ml_input_checker(x)
   assertthat::assert_that(min(x) >= 0)
   rate = 1/mean(x)
   object = c(rate = rate)
