+++ tests/testthat/test_utility.R
@@ -13,3 +13,10 @@ expect_equal(listmerge(x, y, type = "merge")$b, 7)
 expect_null(listmerge(x, y, type = "template")$f, NULL)
 expect_equal(listmerge(x, NULL, type = "merge"), x)
 expect_equal(listmerge(x, list(h = "b"), type = "merge")$h, "b")
+
+## Input checker
+
+expect_error(ml_input_checker("x"))
+expect_error(ml_input_checker(lm))
+expect_error(ml_input_checker(replicate(3, 1:3)))
+expect_true(ml_input_checker(1:3))
