+++ R/mlpareto.R
@@ -23,6 +23,7 @@
 
 mlpareto = function(x, na.rm = FALSE) {
   if(na.rm) x = x[!is.na(x)] else assertthat::assert_that(!anyNA(x))
+  ml_input_checker(x)
   assertthat::assert_that(min(x) > 0)
 
   M = mean(log(x))
