+++ R/mlinvweibull.R
@@ -26,6 +26,7 @@
 
 mlinvweibull = function(x, na.rm = FALSE) {
   if(na.rm) x = x[!is.na(x)] else assertthat::assert_that(!anyNA(x))
+  ml_input_checker(x)
   assertthat::assert_that(min(x) > 0)
 
   y = 1/x
