### 2019-11-12 a818e06
:recycle: Cleaning up data_raw.
