### 2019-11-12 4f945c9
Removed superfluous comments.
