### 2019-11-16 #1863 @MaaniBeigy
Dear @arfon,
Greetings,

I reviewed the **univariateML** package and its software paper. It is an important contribution to the field of statistical programming and has a high quality. My comments are as follow:

1. Although the author has described an MIT license and provided an MIT `LICENSE.md` file, the  LICENSE plain-text file is not correctly an MIT license. I believe, the author should correct the plain-text LICENSE file according to [OSI](https://opensource.org/licenses/alphabetical) and Rbuildignore it (i.e., `usethis::use_build_ignore("LICENSE")` or adding `^LICENSE$` to the `.Rbuildignore` file).

2. The `Date` field in the `DESCRIPTION` file is missing. It causes the following **warning** when a user calls `citation("univariateML")`:

> Warning message:
> In citation("univariateML") :
> no date field in DESCRIPTION file of package ‘univariateML’

- It is better to add something like `Date: 2019-11-06` to the  `DESCRIPTION` file.

3. Although there are clear community guidelines in `README`, labeled as **How to Contribute or Get Help** and in the **Adding-New-Densities** _wiki_, I recommend using a  `CONTRIBUTING.md`.

Please let me know if there are any questions/concerns,
Best regards,
Maani,
