+++ R/mllogis.R
@@ -17,33 +17,37 @@
 #'     \item{\code{support}}{The support of the density.}
 #'     \item{\code{n}}{The number of observations.}
 #'     \item{\code{call}}{The call as captured my \code{match.call}}
-#' @examples mllogis(precip)
+#' @examples
+#' mllogis(precip)
 #' @seealso \link[stats]{Logistic} for the Logistic density, \link[stats]{nlm} for the
 #'   optimizer this function uses.
 #' @references Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995) Continuous Univariate Distributions, Volume 2, Chapter 23. Wiley, New York.
 #' @export
 
-mllogis = function(x, na.rm = FALSE) {
-  if(na.rm) x = x[!is.na(x)] else assertthat::assert_that(!anyNA(x))
+mllogis <- function(x, na.rm = FALSE) {
+  if (na.rm) x <- x[!is.na(x)] else assertthat::assert_that(!anyNA(x))
   ml_input_checker(x)
 
-  m = stats::median(x)
-  mad = stats::median(abs(x - m))
-  start = c(m, log(mad))
+  m <- stats::median(x)
+  mad <- stats::median(abs(x - m))
+  start <- c(m, log(mad))
 
-  f = function(p) -sum(stats::dlogis(x, p[1], exp(p[2]), log = TRUE))
-  values = suppressWarnings(stats::nlm(f = f,
-                      p = start))
+  f <- function(p) -sum(stats::dlogis(x, p[1], exp(p[2]), log = TRUE))
+  values <- suppressWarnings(stats::nlm(
+    f = f,
+    p = start
+  ))
 
-  object = c(location = values$estimate[1],
-             scale = exp(values$estimate[2]))
-  class(object) = "univariateML"
-  attr(object, "model") = "Logistic"
-  attr(object, "density") = "stats::dlogis"
-  attr(object, "logLik") = -values$minimum
-  attr(object, "support") = c(-Inf , Inf)
-  attr(object, "n") = length(x)
-  attr(object, "call") = match.call()
+  object <- c(
+    location = values$estimate[1],
+    scale = exp(values$estimate[2])
+  )
+  class(object) <- "univariateML"
+  attr(object, "model") <- "Logistic"
+  attr(object, "density") <- "stats::dlogis"
+  attr(object, "logLik") <- -values$minimum
+  attr(object, "support") <- c(-Inf, Inf)
+  attr(object, "n") <- length(x)
+  attr(object, "call") <- match.call()
   object
-
 }
