+++ R/helpers.R
@@ -6,16 +6,15 @@
 #' @return A string
 #' @keywords internal
 
-univariateML_to_string = function(obj, type = c("d", "p", "q", "r", "ml")){
-  type = match.arg(type)
-  if(type %in% c("d", "p", "q", "r")) {
-    strings = strsplit(attr(obj, "density"), "::")[[1]]
-    substring(strings[2], first = 1, last = 1) = type
+univariateML_to_string <- function(obj, type = c("d", "p", "q", "r", "ml")) {
+  type <- match.arg(type)
+  if (type %in% c("d", "p", "q", "r")) {
+    strings <- strsplit(attr(obj, "density"), "::")[[1]]
+    substring(strings[2], first = 1, last = 1) <- type
     paste0(strings[1], "::", strings[2])
   } else {
     paste0(type, substring(strsplit(attr(obj, "density"), "::")[[1]][2], 2))
   }
-
 }
 
 #' Transform a univariateML object to a quantile, CDF, density or random
@@ -26,12 +25,12 @@ univariateML_to_string = function(obj, type = c("d", "p", "q", "r", "ml")){
 #' @return A string
 #' @keywords internal
 
-univariateML_to_function = function(obj, type = c("d", "p", "q", "r", "ml")){
-  type = match.arg(type)
-  string = univariateML_to_string(obj, type = type)
-  fun = eval(parse(text = string))
-  if(type %in% c("d", "p", "q", "r")) {
-    for(i in 1:length(obj)) formals(fun)[[names(obj)[i]]] = unname(obj[i])
+univariateML_to_function <- function(obj, type = c("d", "p", "q", "r", "ml")) {
+  type <- match.arg(type)
+  string <- univariateML_to_string(obj, type = type)
+  fun <- eval(parse(text = string))
+  if (type %in% c("d", "p", "q", "r")) {
+    for (i in 1:length(obj)) formals(fun)[[names(obj)[i]]] <- unname(obj[i])
   }
   fun
 }
@@ -41,18 +40,21 @@ univariateML_to_function = function(obj, type = c("d", "p", "q", "r", "ml")){
 #' @param obj Function to apply.
 #' @keywords internal
 
-to_univariateML = function(y, obj) {
-
-  if(inherits(obj, "univariateML")) {
-    obj = obj
+to_univariateML <- function(y, obj) {
+  if (inherits(obj, "univariateML")) {
+    obj <- obj
   } else {
-    msg = "obj must be either a functiona returning a univariateML object or an univariateML object."
-    tryCatch({obj = obj(y)}, error = function(cond) stop(msg))
+    msg <- "obj must be either a functiona returning a univariateML object or an univariateML object."
+    tryCatch(
+      {
+        obj <- obj(y)
+      },
+      error = function(cond) stop(msg)
+    )
     stopifnot(inherits(obj, "univariateML"))
   }
 
   obj
-
 }
 
 
@@ -64,11 +66,11 @@ to_univariateML = function(y, obj) {
 #' @param ... Arguments passed to \code{plot} or \code{points} down the line.
 #' @keywords internal
 
-ppqq_wrangler = function(y, obj, datax, pp, ...) {
+ppqq_wrangler <- function(y, obj, datax, pp, ...) {
 
   ## Nas are removed by default in this function, following qqplot.
 
-  y = y[!is.na(y)]
+  y <- y[!is.na(y)]
 
   ## Error message straight out of stats::qqplot.
   if (0 == length(y)) stop("y is empty or has only NAs")
@@ -76,55 +78,56 @@ ppqq_wrangler = function(y, obj, datax, pp, ...) {
[…]
