### 2019-11-17 6344c02
:hammer: Linting bootstrap.R
