+++ R/bootstrap.R
@@ -1,26 +1,29 @@
 #' Parametric Bootstrap on Distributions Fitted with Maximum Likelihood
 #'
-#' The parametric bootstrap is a resampling technique using random variates from
-#'    a known parametric distribution. In this function the distribution of the
-#'    random variates is completely determined by the \code{unvariateML} object
-#'    \code{object}.
+#' The parametric bootstrap is a resampling technique using random variates
+#'    from a known parametric distribution. In this function the distribution
+#'    of the random variates is completely determined by the \code{unvariateML}
+#'    object \code{object}.
 #'
-#' For each bootstrap iteration a maximum likelihood estimate is calculated using
-#'    the \code{ml***} function specified by \code{object}. The resulting numeric
-#'    vector is then passed to \code{map}. The values returned by \code{map} is collected in
-#'    an array and the \code{reducer} is called on each row of the array.
+#' For each bootstrap iteration a maximum likelihood estimate is calculated
+#'    using the \code{ml***} function specified by \code{object}. The resulting
+#'    numeric vector is then passed to \code{map}. The values returned by
+#'    \code{map} is collected in an array and the \code{reducer} is called on
+#'    each row of the array.
 #'
-#' By default the \code{map} function is the identity and the default \code{reducer} is
-#'    the quantile function taking the argument \code{probs}, which defaults to \code{c(0.025, 0.975)}.
-#'    This corresponds to a 95\% basic percentile confidence interval and is also reported
-#'    by \code{\link{confint}}
+#' By default the \code{map} function is the identity and the default
+#'    \code{reducer} is the quantile function taking the argument \code{probs},
+#'    which defaults to \code{c(0.025, 0.975)}. This corresponds to a 95\%
+#'    basic percentile confidence interval and is also reported by
+#'    \code{\link{confint}}
 #'
-#' \emph{Note:} The default confidence intervals are percentile intervals, not empirical intervals.
-#'    These confidence intervals will in some cases have poor coverage as they are not studentized,
-#'    see e.g. Carpenter, J., & Bithell, J. (2000).
+#' \emph{Note:} The default confidence intervals are percentile intervals,
+#'    not empirical intervals. These confidence intervals will in some cases
+#'    have poor coverage as they are not studentized, see e.g. Carpenter,
+#'    J., & Bithell, J. (2000).
 #'
 #' @param object A \code{univariateML} object.
-#' @param Nreps Positive integer. The number of bootstrap samples.
+#' @param reps Positive integer. The number of bootstrap samples.
 #' @param map A function of the parameters of the \code{univariateML} object.
 #'   Defaults to the identity.
 #' @param reducer A reducer function. Defaults to \code{stats::quantile} with
@@ -29,9 +32,13 @@
 #' @return The transposed map-reduced bootstrap samples.
 #' @export
 #' @references
-#'   Efron, B., & Tibshirani, R. J. (1994). An introduction to the bootstrap. CRC press.
+#'   Efron, B., & Tibshirani, R. J. (1994). An introduction to the bootstrap.
+#'   CRC press.
 #'
-#'   Carpenter, J., & Bithell, J. (2000). Bootstrap confidence intervals: when, which, what? A practical guide for medical statisticians. Statistics in medicine, 19(9), 1141-1164.
+#'   Carpenter, J., & Bithell, J. (2000). Bootstrap confidence intervals:
+#'   when, which, what? A practical guide for medical statisticians.
+#'   Statistics in medicine, 19(9), 1141-1164.
+#'
[…]
