+++ tests/testthat/test_bootstrap.R
@@ -9,7 +9,7 @@ obj2 <- mlexp(x)
 # Different reducer and map.
 set.seed(313)
 expect_equal(c(bootstrapml(obj,
-  Nreps = 2,
+  reps = 2,
   map = function(x) x[1] / x[2],
   reducer = mean
 )),
@@ -20,7 +20,7 @@ tolerance = 1e-5
 # Different reducer and map.
 set.seed(313)
 expect_equal(
-  object = unname(c(bootstrapml(obj, Nreps = 2, reducer = mean))),
+  object = unname(c(bootstrapml(obj, reps = 2, reducer = mean))),
   expected = c(4.333538, 10.581035),
   tolerance = 1e-5
 )
@@ -28,7 +28,7 @@ expect_equal(
 # Common reducer
 set.seed(313)
 expect_equal(
-  object = unname(c(bootstrapml(obj, Nreps = 2))),
+  object = unname(c(bootstrapml(obj, reps = 2))),
   expected = c(2.369753, 5.001551, 6.297323, 16.160518),
   tolerance = 1e-5
 )
@@ -36,14 +36,14 @@ expect_equal(
 # Keeping Names
 set.seed(313)
 expect_equal(
-  object = colnames(bootstrapml(obj, Nreps = 2, reducer = mean)),
+  object = colnames(bootstrapml(obj, reps = 2, reducer = mean)),
   expected = c("shape", "rate")
 )
 
 # Common reducer, one-dim
 set.seed(313)
 expect_equal(
-  object = unname(c(bootstrapml(obj2, Nreps = 2))),
+  object = unname(c(bootstrapml(obj2, reps = 2))),
   expected = c(1.967853, 3.108297),
   tolerance = 1e-5
 )
