### 2019-11-17 24fbdca
seq_along change
