### 2019-11-20 #1863 @JonasMoss
Thanks @MaaniBeigy

1. The way the license is done right now is similar to e.g. [dplyr](https://github.com/tidyverse/dplyr). I'm not sure I can change it since CRAN has asked me to do it in this way (i.e. two lines on `LICENSE` and a `LICENSE.md`) before.

2. Thanks. I think it should be left as it is though, see https://github.com/r-lib/devtools/issues/1327 for an explanation for the warning (which only happens with `devtools::install_github`, not `install.packages`) and https://github.com/MangoTheCat/goodpractice for why it is considered best practice not to have a `Date` field in the description. I'll make sure to make an `INST/CITATION` file to this paper once it has a DOI.

3. I'll add a `CONTRIBUTING.md` based on the one in your `cvcqv` if that's alright. But first I will make the source code tidyverse-style compliant.
