### 2019-11-20 00edf86
:hammer: Linting helpers.R
