### 2019-11-20 #1863 @MaaniBeigy
Hello @JonasMoss ,
Thank you,

1. Yes, it is true and personally acceptable to me. Just, I don't know whether this format is compatible with the criteria of JOSS, "Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?". The two-line format which you have used (like `dplyr`) seems not to be an OSI approved license format. I believe @arfon may help us with that.

2. It is great  :raised_hands: . I did not know the `date` field may also be problematic  :anguished: .

3.  That's alright :+1: . I think having a `CONTRIBUTING.md` is more stylish :sunglasses: , and also helpful for packages which may encourage many to contribute. I believe `univariateML` will draw the attention of statistical programmers to add more distributions, so a well-written easy guide like `CONTRIBUTING.md` facilitates such contributions. However, it is only a suggestion.

Good luck,
Regards,
Maani
