### 2019-11-20 #1863 @arfon
@JonasMoss - please move the contents of the `License.md` file to `LICENSE` too.

The current `LICENSE` file format is somewhat confusing and potentially clashes with the `License.md` contents (so let's just make them the same).

See [tidyverse](https://github.com/tidyverse/tidyverse) as an example of what to do here.
