### 2019-11-20 #1863 @JonasMoss
@arfon @MaaniBeigy I have no problem changing the content of `LICENSE` for the paper release. But I will have to change it back when submitting the package to CRAN, as the CRAN policy demands `LICENSE` only contains two lines when the MIT license is specified in DESCRIPTION.

The CRAN policy is here: https://cran.r-project.org/web/licenses/MIT

I guess the point is that when you visit the CRAN site for a package such as [dplyr](https://cran.r-project.org/web/packages/dplyr/index.html) the usual MIT license is already hosted by CRAN and the LICENSE file shouldn't overlap with it.

Their format is different for the GPL (used in e.g. tidyverse) for some reason.

This policy is enforced: 14 days ago CRAN asked me to change the licensing format when I submitted a package with the format you request:

> > License: MIT + file LICENSE
>
> The MIT license file should only contain the two lines:
>
> YEAR:
> COPYRIGHT HOLDER:
>
> Please fix and resubmit.

Please understand that I **do not** want to make a big deal out this and have no stake in it. (If anything I don't understand why CRAN insists the LICENSE file should contain only these two lines and it's definitely confusing.) But there seems to be a conflict between CRAN's policies and JOSS' policies here.
