+++ R/data-abalone.R
@@ -7,7 +7,8 @@
 #'
 #' @format A \link[tibble]{tibble} with 4,177 observations and 9 variables:
 #'   \describe{
-#'     \item{sex}{Sex of the abalone, \code{F} is female, \code{M} male, and \code{I} infant.}
+#'     \item{sex}{Sex of the abalone, \code{F} is female, \code{M} male, and
+#'                \code{I} infant.}
 #'     \item{length}{Longest shell measurement.}
 #'     \item{diameter}{Diameter perpendicular to length.}
 #'     \item{height}{Height with with meat in shell.}
@@ -18,9 +19,13 @@
 #'     \item{rings}{+1.5 gives the age in years.}
 #'   }
 #'
-#' @source Dua, D. and Graff, C. (2019). UCI Machine Learning Repository \url{http://archive.ics.uci.edu/ml}. Irvine, CA: University of California, School of Information and Computer Science.
+#' @source Dua, D. and Graff, C. (2019). UCI Machine Learning Repository
+#' \url{http://archive.ics.uci.edu/ml}. Irvine, CA: University of California,
+#' School of Information and Computer Science.
+#'
 #' @references
-#'   Ko, V., Hjort, N. L., & Hobaek Haff, I. (2019). Focused information criteria for copulas. Scandinavian Journal of Statistics.
+#'   Ko, V., Hjort, N. L., & Hobaek Haff, I. (2019). Focused information
+#'   criteria for copulas. Scandinavian Journal of Statistics.
 #' @examples
 #' abalone
 "abalone"
