### 2019-11-21 09bd531
:+1: Changing contribution guide
