+++ README.Rmd
@@ -60,16 +60,12 @@ lines(mlweibull(egypt$age))
 
 ## Documentation
 For an overview of the package and its features see the 
-[overview vignette](https://univariateml.netlify.com/articles/overview.html). For a list of implemented 
-densities see the start of the [details vignette](https://univariateml.netlify.com/articles/distributions.html). 
-For an illustration of how this package can make an otherwise long and laborious process
-much simpler, see the [copula vignette](https://univariateml.netlify.com/articles/copula.html).
+[overview vignette](https://univariateml.netlify.com/articles/overview.html). 
+For a list of implemented densities see the start of the [details vignette](https://univariateml.netlify.com/articles/distributions.html). 
+For an illustration of how this package can make an otherwise long and 
+laborious process much simpler, see the [copula vignette](https://univariateml.netlify.com/articles/copula.html).
 
 ## How to Contribute or Get Help
-If you encounter a bug, have a feature request or need some help, open a [Github issue](https://github.com/JonasMoss/univariateML/issues). 
-If you miss a particular estimator, make an issue explaining why its implementation should be prioritized. 
 
-We encourage you to make a pull request if you wish to make a contribution to this package. 
-If you wish to contribute a new maximum likelihood estimator to this package please read [this page first](https://github.com/JonasMoss/univariateML/wiki/Adding-New-Densities). 
+Please read `CONTRIBUTING.md` for details about how to contribute or get help.
 
-This project follows a [Contributor Code of Conduct](https://www.contributor-covenant.org/version/1/4/code-of-conduct.html).
