### 2019-11-21 3a60719
Update CONTRIBUTING.md
