### 2019-11-21 #1863 @JonasMoss
@MaaniBeigy Hi again! I lightly modified your `CONTRIBUTING.md` in the [`CONTRIBUTING.md`](https://github.com/JonasMoss/univariateML/blob/master/CONTRIBUTING.md) for `univariateML`. I hope this is okay!
