### 2019-11-23 #1863 @MaaniBeigy
Hi @JonasMoss @arfon
Great job,
Thank you for your explanations.

Your `CONTRIBUTING.md` sounds great :+1: . You can also add the contents of [Adding New Densities](https://github.com/JonasMoss/univariateML/wiki/Adding-New-Densities) there. It may be a bit difficult for the future contributors to find them in the wiki.  Again, it is just a suggestion.

Regarding the issue of the plain-text LICENSE file, one may also ask the [R-pkg-devel](https://stat.ethz.ch/mailman/listinfo/r-package-devel) mail group.

Regards,
Maani
