### 2019-11-24 #1863 @vbaliga
Hi @JonasMoss and @arfon,

Thanks for giving me extra time to review this. I'm generally very impressed with `univariateML` and I can see myself using it. I'd say the vast majority of my comments are geared towards enhancing clarity of presentation, especially for anyone new to the package. Checklist above has been updated, with some in-line feedback.

I have a few more comments that don't exactly fit into the JOSS review checklist, so I will paste them here.

Best regards,
Vikram
 🐢

#### Additional comments:

- I suggest inserting a `_` within the names of the maximum likelihood estimator functions. I had some trouble distinguishing function names like `mllnorm()` from `mlnorm()` as I tested things out. Having the `_` could help here, e.g. `ml_lnorm()` vs `ml_norm()` is easier for me to distinguish.
- Any interest in creating a delta AIC function? I.e. take the `data.frame` made by `stats::AIC()`, find the minimum AIC value and calculate the delta of each model's AIC from that? Not sure if this already exists elsewhere, but I assume not since you custom-wrote a few lines in your vignettes. But it could be a nice addition to the "Data analysis aids" set of functions that basically supersedes `stats::AIC()` by giving the AICs and also deltas (and maybe AICcs if the user prefers?)
- I also have a suggestion that I'll leave entirely optional at this point: You might give some thought to the 'ecosystem' approach of integrating `univariateML` with your previous `kdensity` package. Can objects made from one package be used by the other in a convenient way? If so, it might be nice to have a 4th vignette demonstrating this and/or some language in the JOSS paper.

#### Some relatively minor spelling/grammatical/formatting things:

- Within the Overview vignette, the first line does not resolve into a bulleted list (at least on the rendered Netlify site)
- Also in the Overview, section "Comparing Many Models with AIC": please revise the line "The generic in `R` can take multiple models, and the lower the the better" so that it's clearer to your reader that the second half of the sentence refers specifically to AIC values.
- Also in the Overview vignette, section "Confidence Intervals", sentence 2: Change "Do do" to "To do"
- In the "Copula Modeling" vignette: "...the copula is know,..." should be "...the copula is known,..."
- `spelling::spell_check_package()` yields mostly false positives, but there are a couple items (e.g. "similarily" in `ProbabilityPlots.Rd:68`) that merit correction
