### 2019-11-25 34a9195
:books: Updated CONTRIBUTING.md
