### 2019-11-25 cf567d4
JOSS Review Spelling error.
