+++ R/probability_plots.R
@@ -13,7 +13,7 @@
 #' an existing plot.
 #'
 #' \code{ppmlplot}, \code{ppmlline}, and \code{ppmlpoints} produce
-#' probability-probability plots (or P-P plots). They behave similarily to the
+#' probability-probability plots (or P-P plots). They behave similarly to the
 #' quantile-quantile plot functions.
 #'
 #' This function is modeled after \link[stats]{qqnorm}.
