### 2019-11-25 5ab6104
:bird: CRAN demands.
I added these to pass CRAN checking with incoming  = TRUE. This functions are imported, but by eval/call.
