+++ R/imports.R
@@ -1,0 +1,7 @@
+# These imports are made to pass a CRAN note. The imports are not redundant,
+# but are used only through eval-call in the code, which is invisible to the
+# CRAN checks.
+
+invisible(extraDistr::dbetapr)
+invisible(logitnorm::dlogitnorm)
+invisible(tibble::tibble)
