### 2019-11-25 f485831
JOSS Review: Documenting speed.
