+++ README.md
@@ -20,8 +20,10 @@ estimation for a
 [selection](https://univariateml.netlify.com/articles/distributions.html)
 of parametric univariate densities. In addition to basic estimation
 capabilities, this package support visualization through `plot` and
-`qqmlplot`, model selection by `AIC` and `BIC`, and confidence sets
-through the parametric bootstrap with `bootstrapml`.
+`qqmlplot`, model selection by `AIC` and `BIC`, confidence sets through
+the parametric bootstrap with `bootstrapml`, and convenience functions
+such as the density, distribution function, quantile function, and
+random sampling at the estimated distribution parameters.
 
 ## Installation
 
@@ -54,6 +56,31 @@ lines(mlweibull(egypt$age))
 ```
 
 <img src="man/figures/README-weibull_plot-1.png" width="750px" />
+
+## Implementations
+
+Analytic formulae for the maximum likelihood estimates are used whenever
+they exist. Most `ml***` functions without analytic solutions have a
+custom made Newton-Raphson solver. These can be much faster than the
+naïve solvers using `nlm` or `optim`. For example, `mlbeta` has a large
+speedup.
+
+``` r
+# install.packages("microbenchmark")
+set.seed(313)
+x <- rbeta(500, 2, 7)
+
+microbenchmark::microbenchmark(
+  univariateML = univariateML::mlbeta(x),
+  naive = nlm(function(p) -sum(dbeta(x, p[1], p[2], log = TRUE)), p = c(1, 1)))
+#> Unit: microseconds
+#>          expr     min       lq      mean   median       uq     max neval
+#>  univariateML   864.2  1102.90  1305.743  1192.30  1340.95  6007.9   100
+#>         naive 36398.0 39097.35 41108.804 40530.15 43035.50 50693.3   100
+```
+
+The maximum likelihood estimators in this package have all been subject
+to testing, see the `tests` folder for details.
 
 ## Documentation
 
