### 2019-11-25 0e157fe
:books: JOSS Review Vignettes
