### 2019-11-25 5587967
:book: JOSS Review, more info on speedup.
