### 2019-11-25 3627bf9
:book: Added 'copula' and 'dplyr' conditional install; fixed vignette index title.
