+++ R/data-egypt.R
@@ -7,8 +7,7 @@
 #'    Claeskens & Hjort (2008) and the data is based on their transcription.
 #'
 #' @seealso The source of the data is
-#' \url{https://feb.kuleuven.be/public/u0043181/modelselection/
-#' datasets/egyptlives_data.txt}
+#' \url{https://feb.kuleuven.be/public/u0043181/modelselection/datasets/egyptlives_data.txt}
 #'
 #' @format A \link[tibble]{tibble} with 141 observations and 2 variables:
 #'   \describe{
