+++ vignettes/copula.Rmd
@@ -9,8 +9,13 @@ bibliography: copula.bib
 ---
 
 ```{R depends, echo = FALSE}
-if (!require("copula")) install.packages("copula", dependencies = TRUE)
-if (!require("dplyr")) install.packages("dplyr", dependencies = TRUE)
+if (!require("copula")) install.packages("copula", 
+                                         dependencies = TRUE,
+                                         repos = "http://cran.us.r-project.org")
+
+if (!require("dplyr")) install.packages("dplyr", 
+                                        dependencies = TRUE,
+                                        repos = "http://cran.us.r-project.org")
 ```
 
 This vignette shows how to model multivariate distributions with 
