### 2019-11-25 9fa16fc
Update egypt.Rd
