### 2019-11-25 c9a80d7
JOSS Update: Reordered Vignettes.
