+++ README.Rmd
@@ -83,7 +83,8 @@ The maximum likelihood estimators in this package have all been subject to
 testing, see the `tests` folder for details.
 
 ## Documentation
-For an overview of the package and its features see the 
+This package has a [website](https://univariateml.netlify.com/index.html)). For 
+an overview of the package and its features see the 
 [overview vignette](https://univariateml.netlify.com/articles/overview.html). 
 For a list of implemented densities see the start of the [details vignette](https://univariateml.netlify.com/articles/distributions.html). 
 For an illustration of how this package can make an otherwise long and 
