### 2019-11-25 bdeaf8f
Update paper.rmd
