### 2019-11-25 7db514d
:books: Paper update.
* Fixing referecnes;
* Minor style change;
* A spelling error.
