+++ paper/paper.md
@@ -18,17 +18,14 @@ bibliography: paper.bib
 
 # Summary
 
-`univariateML` is an R (R Core Team 2019) package for doing univariate
-maximum likelihood estimation (Cam 1990). It supports more than 20
-densities, the most popular generic functions such as `plot`, `AIC`, and
-`confint`, and a simple parametric bootstrap (Efron and Tibshirani 1994)
-interface.
+`univariateML` is an R [@r] package for univariate maximum likelihood estimation [@lecam1990ml]. 
+It supports more than 20 densities, the most popular generic functions such as `plot`, `AIC`, and `confint`, and a simple parametric bootstrap [@efron1994introduction] interface.
 
-When looking at univariate data it is natural to ask if there is a known
-parametric density that fits the data well. The following example uses
-the `egypt` (Pearson 1902) data set included in the package and a plot
-of the Weibull and Gamma densities (Johnson, Kotz, and Balakrishnan
-1995, Chapter 17 & 21).
+When looking at univariate data it is natural to ask if there is a known 
+parametric density that fits the data well. The following example uses the 
+`egypt` [@pearson1902egypt] data set included in the package and a plot of the Weibull and Gamma
+densities [@johnson1970continuous, Chapter 17 & 21]. 
+
 
 ``` r
 # install.packages("univariateML")
@@ -40,9 +37,8 @@ lines(mlgamma(egypt$age), col = "red")  # Plots a Gamma fit.
 
 ![](paper_files/figure-gfm/figure-1.png)<!-- -->
 
-Another natural question to ask is which among several models fits the
-data best. This can be done using tools of model selection such as the
-`AIC` (Akaike 1998).
+A natural question to ask is which among several models fits the data best.
+This can be done using tools of model selection such as the `AIC` [@akaike1998information].
 
 ``` r
 AIC(mlweibull(egypt$age),
@@ -53,39 +49,37 @@ AIC(mlweibull(egypt$age),
     ## mlweibull(egypt$age)  2 1230.229
     ## mlgamma(egypt$age)    2 1234.772
 
-Problems involving estimation of univariate densities are common in
-statistics. Estimation of univariate densities is used in for instance
-exploratory data analysis, in the estimation of copulas (Ko, Hjort, and
-Hobæk Haff 2019), as parametric starts in density estimation (Hjort and
-Glad 1995; Moss and Tveten 2019), and is of interest in and of itself.
 
-`univariateML` exists to simplify the whole process of doing inference
-with univariate densities. For most densities implemented in `R` the
-maximum likelihood estimates can easily be computed using numerical
-optimization functions such as `stats::nlm` and `stats::optim` on the
-negative log-likelihood, but there are three problems with this solution
-strategy:
+Problems involving estimation of univariate densities are common in statistics. 
+Estimation of univariate densities is used in for instance exploratory data analysis, 
+in the estimation of copulas [@ko2019focused],
+as parametric starts in density estimation [@hjort_glad_1995; @moss2019kdensity], 
+and is of interest in and of itself. 
 
-1.  It takes much time to program, especially if you wish to try out
-    many densities and wish to do make for instance `qqplot` or density
-    plots.
-2.  It is bug prone.
-3.  The estimation itself can be slow when `$n$` is large. The time lost
-    quickly adds up when doing the parametric bootstrap or another
-    procedure requiring repeated calls to the estimating function.
+`univariateML` exists to simplify the whole process of doing inference with univariate densities.
+For most densities implemented in `R` the maximum likelihood estimates can easily
+be computed using numerical optimization functions such as `stats::nlm` and
+`stats::optim` on the negative log-likelihood, but there are three problems 
+with this solution strategy:
+
[…]
