+++ paper/paper.rmd
@@ -20,7 +20,7 @@ bibliography: paper.bib
 
 # Summary
 
-`univariateML` is an R [@r] package for doing univariate maximum likelihood estimation [@lecam1990ml]. 
+`univariateML` is an R [@r] package for univariate maximum likelihood estimation [@lecam1990ml]. 
 It supports more than 20 densities, the most popular generic functions such as `plot`, `AIC`, and `confint`, and a simple parametric bootstrap [@efron1994introduction] interface.
 
 When looking at univariate data it is natural to ask if there is a known 
@@ -36,7 +36,7 @@ lines(mlweibull(egypt$age)) # Plots a Weibull fit.
 lines(mlgamma(egypt$age), col = "red")  # Plots a Gamma fit.
 ```
 
-Another natural question to ask is which among several models fits the data best.
+A natural question to ask is which among several models fits the data best.
 This can be done using tools of model selection such as the `AIC` [@akaike1998information].
 
 ```{r, AIC}
@@ -56,8 +56,8 @@ be computed using numerical optimization functions such as `stats::nlm` and
 `stats::optim` on the negative log-likelihood, but there are three problems 
 with this solution strategy:
 
-1. It takes much time to program, especially if you wish to try out many densities and
-   wish to do make for instance `qqplot` or density plots.
+1. It takes much time to program, especially if we wish to try out many densities and
+   wish to make for instance `qqplot` or density plots.
 2. It is bug prone.
 3. The estimation itself can be slow when `$n$` is large. The time lost quickly adds up
    when doing the parametric bootstrap or another procedure requiring repeated calls to
