+++ README.Rmd
@@ -27,7 +27,7 @@ set.seed(313)
 ```
 
 ## Overview
-An `R`-package for fast, easy, and reliable maximum likelihood estimation for a 
+[`univariateML`](https://univariateml.netlify.com/index.html) is an `R`-package for fast, easy, and reliable maximum likelihood estimation for a 
 [selection](https://univariateml.netlify.com/articles/distributions.html) of parametric univariate densities. In addition to basic estimation capabilities,
 this package support visualization through `plot` and `qqmlplot`, model selection
 by `AIC` and `BIC`, confidence sets through the parametric bootstrap with
@@ -83,8 +83,7 @@ The maximum likelihood estimators in this package have all been subject to
 testing, see the `tests` folder for details.
 
 ## Documentation
-This package has a [website](https://univariateml.netlify.com/index.html). For 
-an overview of the package and its features see the 
+For an overview of the package and its features see the 
 [overview vignette](https://univariateml.netlify.com/articles/overview.html). 
 For a list of implemented densities see the start of the [details vignette](https://univariateml.netlify.com/articles/distributions.html). 
 For an illustration of how this package can make an otherwise long and 
