+++ paper/paper.rmd
@@ -59,7 +59,7 @@ with this solution strategy:
 1. It takes much time to program, especially if we wish to try out many densities and
    wish to make for instance `qqplot` or density plots.
 2. It is bug prone.
-3. The estimation itself can be slow when `$n$` is large. The time lost quickly adds up
+3. The estimation itself can be slow when when the sample size. The time lost quickly adds up
    when doing the parametric bootstrap or another procedure requiring repeated calls to
    the estimating function.
 
