+++ paper/paper.bib
@@ -1,4 +1,12 @@
-﻿@article{hjort_glad_1995,
+﻿@Manual{Rfast,
+    title = {Rfast: A Collection of Efficient and Extremely Fast R Functions},
+    author = {Manos Papadakis and Michail Tsagris and Marios Dimitriadis and Stefanos Fafalios and Ioannis Tsamardinos and Matteo Fasiolo and Giorgos Borboudakis and John Burkardt and Changliang Zou and Kleanthi Lakiotaki and Christina Chatzipantsiou.},
+    year = {2019},
+    note = {R package version 1.9.6},
+    url = {https://CRAN.R-project.org/package=Rfast},
+}
+
+@article{hjort_glad_1995,
 	title={Nonparametric density estimation with a parametric start},
 	author={Hjort, Nils Lid and Glad, Ingrid K},
 	journal={The Annals of Statistics},
