+++ NAMESPACE
@@ -27,6 +27,7 @@ export(mllnorm)
 export(mllogis)
 export(mllogitnorm)
 export(mllomax)
+export(mlnaka)
 export(mlnorm)
 export(mlpareto)
 export(mlpower)
