+++ R/data-abalone.R
@@ -2,13 +2,13 @@
 #'
 #' Physical measurements of 4177 abalones, a species of sea snail.
 #'
-#' See the web page \url{https://archive.ics.uci.edu/ml/datasets/Abalone} for
+#' See the web page <https://archive.ics.uci.edu/ml/datasets/Abalone> for
 #'    more information about the data set.
 #'
-#' @format A \link[tibble]{tibble} with 4,177 observations and 9 variables:
+#' @format A [tibble][tibble::tibble] with 4,177 observations and 9 variables:
 #'   \describe{
-#'     \item{sex}{Sex of the abalone, \code{F} is female, \code{M} male, and
-#'                \code{I} infant.}
+#'     \item{sex}{Sex of the abalone, `F` is female, `M` male, and
+#'                `I` infant.}
 #'     \item{length}{Longest shell measurement.}
 #'     \item{diameter}{Diameter perpendicular to length.}
 #'     \item{height}{Height with with meat in shell.}
@@ -20,7 +20,7 @@
 #'   }
 #'
 #' @source Dua, D. and Graff, C. (2019). UCI Machine Learning Repository
-#' \url{http://archive.ics.uci.edu/ml}. Irvine, CA: University of California,
+#' <http://archive.ics.uci.edu/ml>. Irvine, CA: University of California,
 #' School of Information and Computer Science.
 #'
 #' @references
