+++ R/mlrayleigh.R
@@ -1,23 +1,23 @@
 #' Rayleigh distribution maximum likelihood estimation
 #'
-#' Calculates the \code{sigma} parameter as the square root of half the
+#' Calculates the `sigma` parameter as the square root of half the
 #'    empirical second moment.
 #'
-#' For the density function of the Rayleigh distribution see \link[extraDistr]{Rayleigh}.
+#' For the density function of the Rayleigh distribution see [Rayleigh][extraDistr::Rayleigh].
 #'
 #' @param x a (non-empty) numeric vector of data values.
 #' @param na.rm logical. Should missing values be removed?
-#' @return \code{mlrayleigh} returns an object of \link[base]{class} \code{univariateML}. This
-#'    is a named numeric vector with maximum likelihood estimates for \code{sigma} and the following attributes:
-#'     \item{\code{model}}{The name of the model.}
-#'     \item{\code{density}}{The density associated with the estimates.}
-#'     \item{\code{logLik}}{The loglikelihood at the maximum.}
-#'     \item{\code{support}}{The support of the density.}
-#'     \item{\code{n}}{The number of observations.}
-#'     \item{\code{call}}{The call as captured my \code{match.call}}
+#' @return `mlrayleigh` returns an object of [class][base::class] `univariateML`. This
+#'    is a named numeric vector with maximum likelihood estimates for `sigma` and the following attributes:
+#'     \item{`model`}{The name of the model.}
+#'     \item{`density`}{The density associated with the estimates.}
+#'     \item{`logLik`}{The loglikelihood at the maximum.}
+#'     \item{`support`}{The support of the density.}
+#'     \item{`n`}{The number of observations.}
+#'     \item{`call`}{The call as captured my `match.call`}
 #' @examples
 #' mlrayleigh(precip)
-#' @seealso \link[extraDistr]{Rayleigh} for the Rayleigh density.
+#' @seealso [Rayleigh][extraDistr::Rayleigh] for the Rayleigh density.
 #' @references Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995) Continuous Univariate Distributions, Volume 1, Chapter 18. Wiley, New York.
 #' @export
 
