+++ R/mlinvgamma.R
@@ -2,7 +2,7 @@
 #'
 #' Transforms the data and uses Newton-Raphson to estimate the parameters of the Gamma distribution.
 #'
-#' For the density function of the Gamma distribution see \link[extraDistr]{InvGamma}.
+#' For the density function of the Gamma distribution see [InvGamma][extraDistr::InvGamma].
 #'
 #' @param x a (non-empty) numeric vector of data values.
 #' @param na.rm logical. Should missing values be removed?
@@ -10,10 +10,10 @@
 #' @param iterlim A positive integer specifying the maximum number of
 #' iterations to be performed before the program is terminated.
 #' @return A named numeric vector with maximum likelihood estimates for
-#' \code{alpha} and \code{beta}.
+#' `alpha` and `beta`.
 #' @examples
 #' mlinvgamma(precip)
-#' @seealso \link[extraDistr]{InvGamma} for the Inverse Gamma density.
+#' @seealso [InvGamma][extraDistr::InvGamma] for the Inverse Gamma density.
 #' @references Choi, S. C, and R. Wette. "Maximum likelihood estimation of the parameters of the gamma distribution and their bias." Technometrics 11.4 (1969): 683-690.
 #' Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995) Continuous Univariate Distributions, Volume 1, Chapter 17. Wiley, New York.
 #' Witkovsky, V. (2001). "Computing the Distribution of a Linear Combination of Inverted Gamma Variables". Kybernetika. 37 (1): 79–90
