+++ R/data-egypt.R
@@ -7,9 +7,9 @@
 #'    Claeskens & Hjort (2008) and the data is based on their transcription.
 #'
 #' @seealso The source of the data is
-#' \url{https://feb.kuleuven.be/public/u0043181/modelselection/datasets/egyptlives_data.txt}
+#' <https://feb.kuleuven.be/public/u0043181/modelselection/datasets/egyptlives_data.txt>
 #'
-#' @format A \link[tibble]{tibble} with 141 observations and 2 variables:
+#' @format A [tibble][tibble::tibble] with 141 observations and 2 variables:
 #'   \describe{
 #'     \item{age}{Age at death.}
 #'     \item{sex}{Sex of deceased; 82 males and 49 females.}

[…]
