### 2019-12-03 3948f2b
Using `...` in mlcauchy, mlexp, and mlgamma.
