+++ R/mlgamma.R
@@ -2,15 +2,18 @@
 #'
 #' Uses Newton-Raphson to estimate the parameters of the Gamma distribution.
 #'
-#' For the density function of the Gamma distribution see [GammaDist][stats::GammaDist].
+#' For the density function of the Gamma distribution see
+#' [GammaDist][stats::GammaDist].
 #'
 #' @param x a (non-empty) numeric vector of data values.
 #' @param na.rm logical. Should missing values be removed?
-#' @param rel.tol Relative accuracy requested.
-#' @param iterlim A positive integer specifying the maximum number of
-#' iterations to be performed before the program is terminated.
-#' @return `mlgamma` returns an object of [class][base::class] `univariateML`. This
-#'    is a named numeric vector with maximum likelihood estimates for `shape` and `rate` and the following attributes:
+#' @param ... `rel.tol` is the relative accuracy requested, defaults
+#'     to `.Machine$double.eps^0.25`. `iterlim` is a positive integer
+#'     specifying the maximum number of iterations to be performed before the
+#'     program is terminated (defaults to `100`).
+#' @return `mlgamma` returns an object of [class][base::class] `univariateML`.
+#'    This is a named numeric vector with maximum likelihood estimates for
+#'    `shape` and `rate` and the following attributes:
 #'     \item{`model`}{The name of the model.}
 #'     \item{`density`}{The density associated with the estimates.}
 #'     \item{`logLik`}{The loglikelihood at the maximum.}
@@ -20,17 +23,26 @@
 #' @examples
 #' mlgamma(precip)
 #' @seealso [GammaDist][stats::GammaDist] for the Gamma density.
-#' @references Choi, S. C, and R. Wette. "Maximum likelihood estimation of the parameters of the gamma distribution and their bias." Technometrics 11.4 (1969): 683-690.
-#' Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995) Continuous Univariate Distributions, Volume 1, Chapter 17. Wiley, New York.
+#' @references Choi, S. C, and R. Wette. "Maximum likelihood estimation
+#' of the parameters of the gamma distribution and their bias."
+#' Technometrics 11.4 (1969): 683-690.
+#'
+#' Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995) Continuous
+#' Univariate Distributions, Volume 1, Chapter 17. Wiley, New York.
 #' @export
 
-mlgamma <- function(x, na.rm = FALSE, rel.tol = .Machine$double.eps^0.25,
-                    iterlim = 100) {
+mlgamma <- function(x, na.rm = FALSE, ...) {
   if (na.rm) x <- x[!is.na(x)] else assertthat::assert_that(!anyNA(x))
   ml_input_checker(x)
   assertthat::assert_that(min(x) > 0)
 
-  rel.tol_str <- deparse(substitute(rel.tol))
+  dots <- list(...)
+
+  rel.tol <- if(!is.null(dots$rel.tol)) dots$rel.tol else
+    .Machine$double.eps^0.25
+
+  iterlim <- if(!is.null(dots$iterlim)) dots$iterlim else
+    100
 
   n <- length(x)
   mean_hat <- mean(x)
@@ -42,7 +54,8 @@ mlgamma <- function(x, na.rm = FALSE, rel.tol = .Machine$double.eps^0.25,
 
   ## The Newton-Raphson steps.
   for (i in 1:iterlim) {
-    shape <- shape0 - (log(shape0) - digamma(shape0) - s)/(1 / shape0 - trigamma(shape0))
+    shape <- shape0 - (log(shape0) - digamma(shape0) - s)/
[…]
