+++ R/mlcauchy.R
@@ -5,26 +5,30 @@
 #'     the edge case where no maximum likelihood estimator exists and error is
 #'     thrown.
 #'
-#' For the density function of the Cauchy distribution see [Cauchy][stats::Cauchy].
+#' For the density function of the Cauchy distribution see
+#'     [Cauchy][stats::Cauchy].
 #'
 #' @param x a (non-empty) numeric vector of data values.
 #' @param na.rm logical. Should missing values be removed?
-#' @return `mlcauchy` returns an object of [class][base::class] `univariateML`. This
-#'    is a named numeric vector with maximum likelihood estimates for `location` and `scale` and the following attributes:
+#' @param ... currently affects nothing.
+#' @return `mlcauchy` returns an object of [class][base::class] `univariateML`.
+#'    This is a named numeric vector with maximum likelihood estimates for
+#'    `location` and `scale` and the following attributes:
 #'     \item{`model`}{The name of the model.}
 #'     \item{`density`}{The density associated with the estimates.}
 #'     \item{`logLik`}{The loglikelihood at the maximum.}
 #'     \item{`support`}{The support of the density.}
 #'     \item{`n`}{The number of observations.}
 #'     \item{`call`}{The call as captured my `match.call`}
-#' @seealso [Cauchy][stats::Cauchy] for the Cauchy density, [nlm][stats::nlm] for the
-#'   optimizer this function uses.
-#' @references  #' @references Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995) Continuous Univariate Distributions, Volume 1, Chapter 16. Wiley, New York.
+#' @seealso [Cauchy][stats::Cauchy] for the Cauchy density, [nlm][stats::nlm]
+#'   for the optimizer this function uses.
+#' @references Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995)
+#' Continuous Univariate Distributions, Volume 1, Chapter 16. Wiley, New York.
 #' @examples
 #' mlcauchy(airquality$Temp)
 #' @export
 
-mlcauchy <- function(x, na.rm = FALSE) {
+mlcauchy <- function(x, na.rm = FALSE, ...) {
   if (na.rm) x <- x[!is.na(x)] else assertthat::assert_that(!anyNA(x))
   ml_input_checker(x)
 
