+++ R/mlinvgauss.R
@@ -4,12 +4,15 @@
 #'     maximum likelihood estimate of `1/shape` is the difference between
 #'     the mean of reciprocals and the reciprocal of the mean.
 #'
-#' For the density function of the Inverse Gamma distribution see [InverseGaussian][actuar::InverseGaussian].
+#' For the density function of the Inverse Gamma distribution see
+#'     [InverseGaussian][actuar::InverseGaussian].
 #'
 #' @param x a (non-empty) numeric vector of data values.
 #' @param na.rm logical. Should missing values be removed?
-#' @return `mlinvgauss` returns an object of [class][base::class] `univariateML`. This
-#'    is a named numeric vector with maximum likelihood estimates for `mean` and `shape` and the following attributes:
+#' @param ... currently affects nothing.
+#' @return `mlinvgauss` returns an object of [class][base::class]
+#'    `univariateML`. This is a named numeric vector with maximum likelihood
+#'    estimates for `mean` and `shape` and the following attributes:
 #'     \item{`model`}{The name of the model.}
 #'     \item{`density`}{The density associated with the estimates.}
 #'     \item{`logLik`}{The loglikelihood at the maximum.}
@@ -18,11 +21,13 @@
 #'     \item{`call`}{The call as captured my `match.call`}
 #' @examples
 #' mlinvgauss(precip)
-#' @seealso [InverseGaussian][actuar::InverseGaussian] for the Inverse Gaussian density.
-#' @references Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995) Continuous Univariate Distributions, Volume 1, Chapter 15. Wiley, New York.
+#' @seealso [InverseGaussian][actuar::InverseGaussian] for the
+#' Inverse Gaussian density.
+#' @references Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995)
+#' Continuous Univariate Distributions, Volume 1, Chapter 15. Wiley, New York.
 #' @export
 
-mlinvgauss <- function(x, na.rm = FALSE) {
+mlinvgauss <- function(x, na.rm = FALSE, ...) {
   if (na.rm) x <- x[!is.na(x)] else assertthat::assert_that(!anyNA(x))
   ml_input_checker(x)
   assertthat::assert_that(min(x) > 0)
@@ -37,7 +42,8 @@ mlinvgauss <- function(x, na.rm = FALSE) {
   class(object) <- "univariateML"
   attr(object, "model") <- "Inverse Gaussian"
   attr(object, "density") <- "actuar::dinvgauss"
-  attr(object, "logLik") <- -n / 2 * (3 * L - log(lambda) + log(2 * pi) + lambda / mu^2 * S)
+  attr(object, "logLik") <-
+    -n / 2 * (3 * L - log(lambda) + log(2 * pi) + lambda / mu^2 * S)
   attr(object, "support") <- c(0, Inf)
   attr(object, "n") <- length(x)
   attr(object, "call") <- match.call()

[…]
