### 2019-12-03 ecb3f28
:hammer: Added test of formals.
