### 2019-12-03 f8b2572
:hammer: Linting tests.
