+++ R/mllgamma.R
@@ -3,16 +3,15 @@
 #' The maximum likelihood estimate of `shapelog` and `ratelog` are calculated
 #'   by calling [mlgamma()] on the transformed data.
 #'
-#' For the density function of the log normal distribution see [Loggamma][actuar::Loggamma].
+#' For the density function of the log normal distribution see
+#'   [Loggamma][actuar::Loggamma].
 #'
 #' @param x a (non-empty) numeric vector of data values.
 #' @param na.rm logical. Should missing values be removed?
-#' @param rel.tol Relative accuracy; passed to `mlgamma`.
-#' @param iterlim A positive integer specifying the maximum number of
-#' iterations to be performed before the program is terminated. Passed to
-#' `mlgamma`.
-#' @return `mllgamma` returns an object of [class][base::class] `univariateML`. This
-#'    is a named numeric vector with maximum likelihood estimates for `shapelog` and `ratelog` and the following attributes:
+#' @param ... passed to [`mlgamma`][mlgamma].
+#' @return `mllgamma` returns an object of [class][base::class] `univariateML`.
+#'    This is a named numeric vector with maximum likelihood estimates for
+#'    `shapelog` and `ratelog` and the following attributes:
 #'     \item{`model`}{The name of the model.}
 #'     \item{`density`}{The density associated with the estimates.}
 #'     \item{`logLik`}{The loglikelihood at the maximum.}
@@ -24,11 +23,11 @@
 #' @seealso [Loggamma][actuar::Loggamma] for the log normal density.
 #' @references Hogg, R. V. and Klugman, S. A. (1984), Loss Distributions, Wiley.
 #'
-#' Dutang, C., Goulet, V., & Pigeon, M. (2008). actuar: An R package for actuarial science. Journal of Statistical Software, 25(7), 1-37.
+#' Dutang, C., Goulet, V., & Pigeon, M. (2008). actuar: An R package for
+#' actuarial science. Journal of Statistical Software, 25(7), 1-37.
 #' @export
 
-mllgamma <- function(x, na.rm = FALSE, rel.tol = .Machine$double.eps^0.25,
-                     iterlim = 100) {
+mllgamma <- function(x, na.rm = FALSE, ...) {
   if (na.rm) x <- x[!is.na(x)] else assertthat::assert_that(!anyNA(x))
   ml_input_checker(x)
   assertthat::assert_that(min(x) > 1)
@@ -36,7 +35,7 @@ mllgamma <- function(x, na.rm = FALSE, rel.tol = .Machine$double.eps^0.25,
   y <- log(x)
   n <- length(x)
 
-  object <- mlgamma(y, na.rm = na.rm, rel.tol = rel.tol, iterlim = iterlim)
+  object <- mlgamma(y, na.rm = na.rm, ...)
   shapelog <- object[1]
   ratelog <- object[2]
   L <- mean(y)
@@ -45,8 +44,9 @@ mllgamma <- function(x, na.rm = FALSE, rel.tol = .Machine$double.eps^0.25,
   attr(object, "model") <- "Loggamma"
   attr(object, "density") <- "actuar::dlgamma"
   attr(object, "support") <- c(1, Inf)
-  attr(object, "logLik") <- unname(n * (shapelog * log(ratelog) - log(gamma(shapelog)) +
-    (shapelog - 1) * K - (ratelog + 1) * L))
+  attr(object, "logLik") <-
+    unname(n * (shapelog * log(ratelog) - log(gamma(shapelog)) +
+      (shapelog - 1) * K - (ratelog + 1) * L))
 
   attr(object, "call") <- match.call()
   object
