+++ man/mlpower.Rd
@@ -4,19 +4,20 @@
 \alias{mlpower}
 \title{Power distribution maximum likelihood estimation}
 \usage{
-mlpower(x, na.rm = FALSE, epsilon = .Machine$double.eps^0.5)
+mlpower(x, na.rm = FALSE, ...)
 }
 \arguments{
 \item{x}{a (non-empty) numeric vector of data values.}
 
 \item{na.rm}{logical. Should missing values be removed?}
 
-\item{epsilon}{Positive number added to \code{max(x)} as an approximation
-to the maximum likelihood.}
+\item{...}{\code{epsilon} is a positive number added to \code{max(x)} as an to the
+maximum likelihood. Defaults to \code{.Machine$double.eps^0.5}.}
 }
 \value{
-\code{mlpower} returns an object of \link[base:class]{class} \code{univariateML}. This
-is a named numeric vector with maximum likelihood estimates for \code{alpha} and \code{beta} and the following attributes:
+\code{mlpower} returns an object of \link[base:class]{class} \code{univariateML}.
+This is a named numeric vector with maximum likelihood estimates for
+\code{alpha} and \code{beta} and the following attributes:
 \item{\code{model}}{The name of the model.}
 \item{\code{density}}{The density associated with the estimates.}
 \item{\code{logLik}}{The loglikelihood at the maximum.}
@@ -30,8 +31,9 @@ The maximum likelihood estimate of \code{alpha} is the maximum of \code{x} +
 \code{beta} is \code{1/(log(alpha)-mean(log(x)))}.
 }
 \details{
-For the density function of the power distribution see \link[extraDistr:PowerDist]{PowerDist}.
-The maximum likelihood estimator of \code{alpha} does not exist, strictly
+For the density function of the power distribution see
+\link[extraDistr:PowerDist]{PowerDist}. The maximum likelihood estimator of
+\code{alpha} does not exist, strictly
 speaking. This is because \code{x} is supported \code{c(0, alpha)} with
 an open endpoint on alpha in the \code{extraDistr} implementation of
 \code{dpower}. If the endpoint was closed, \code{max(x)} would have been
@@ -42,9 +44,10 @@ a possibly user specified \code{epsilon} to \code{max(x)}.
 mlpower(precip)
 }
 \references{
-Arslan, G. "A new characterization of the power distribution." Journal of Computational and Applied Mathematics 260 (2014): 99-102.
+Arslan, G. "A new characterization of the power distribution."
+Journal of Computational and Applied Mathematics 260 (2014): 99-102.
 }
 \seealso{
-\link[extraDistr:PowerDist]{PowerDist} for the power density. \link{Pareto} for
-the closely related Pareto distribution.
+\link[extraDistr:PowerDist]{PowerDist} for the power density. \link{Pareto}
+for the closely related Pareto distribution.
 }
