+++ man/mlrayleigh.Rd
@@ -4,16 +4,19 @@
 \alias{mlrayleigh}
 \title{Rayleigh distribution maximum likelihood estimation}
 \usage{
-mlrayleigh(x, na.rm = FALSE)
+mlrayleigh(x, na.rm = FALSE, ...)
 }
 \arguments{
 \item{x}{a (non-empty) numeric vector of data values.}
 
 \item{na.rm}{logical. Should missing values be removed?}
+
+\item{...}{currently affects nothing.}
 }
 \value{
-\code{mlrayleigh} returns an object of \link[base:class]{class} \code{univariateML}. This
-is a named numeric vector with maximum likelihood estimates for \code{sigma} and the following attributes:
+\code{mlrayleigh} returns an object of \link[base:class]{class}
+\code{univariateML}. This is a named numeric vector with maximum likelihood
+estimates for \code{sigma} and the following attributes:
 \item{\code{model}}{The name of the model.}
 \item{\code{density}}{The density associated with the estimates.}
 \item{\code{logLik}}{The loglikelihood at the maximum.}
@@ -26,13 +29,15 @@ Calculates the \code{sigma} parameter as the square root of half the
 empirical second moment.
 }
 \details{
-For the density function of the Rayleigh distribution see \link[extraDistr:Rayleigh]{Rayleigh}.
+For the density function of the Rayleigh distribution see
+\link[extraDistr:Rayleigh]{Rayleigh}.
 }
 \examples{
 mlrayleigh(precip)
 }
 \references{
-Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995) Continuous Univariate Distributions, Volume 1, Chapter 18. Wiley, New York.
+Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995) Continuous
+Univariate Distributions, Volume 1, Chapter 18. Wiley, New York.
 }
 \seealso{
 \link[extraDistr:Rayleigh]{Rayleigh} for the Rayleigh density.
