### 2019-12-03 17f341f
:book: Paper update.
