+++ README.Rmd
@@ -27,7 +27,8 @@ set.seed(313)
 ```
 
 ## Overview
-[`univariateML`](https://univariateml.netlify.com/index.html) is an `R`-package for fast, easy, and reliable maximum likelihood estimation for a 
+[`univariateML`](https://univariateml.netlify.com/index.html) is an `R`-package for
+user-friendly maximum likelihood estimation of a
 [selection](https://univariateml.netlify.com/articles/distributions.html) of parametric univariate densities. In addition to basic estimation capabilities,
 this package support visualization through `plot` and `qqmlplot`, model selection
 by `AIC` and `BIC`, confidence sets through the parametric bootstrap with
