### 2019-12-03 #1863 @JonasMoss
@MaaniBegi
> Your `CONTRIBUTING.md` sounds great 👍 . You can also add the contents of [Adding New Densities](https://github.com/JonasMoss/univariateML/wiki/Adding-New-Densities) there. It may be a bit difficult for the future contributors to find them in the wiki. Again, it is just a suggestion.

Sounds reasonable. I've changed `CONTRIBUTING.md` accordingly.
