### 2019-12-03 #1863 @JonasMoss
Hi @vbaliga and thanks for your comments!

I've responded to your suggestions below, implementing all except the two under the additional comments header. I hope the paper is satisfactory too.

## Checklist
> #### **Installation:**
> Yes, although I note that using either `devools::check()` or `devtools::install_github("JonasMoss/univariateML", build_vignettes = TRUE)` fails when attempting to create vignettes. This is likely because of how `copula` is needed but not installed & loaded by the "Copula Modeling" vignette. I suggest revising how `copula` is handled within this vignette just so that users who want local vignettes don't have a hard time.

Thanks! I had no idea this could be a problem.  I copied your solution in https://github.com/JonasMoss/univariateML/commit/0c993bd7293c5e961c0de6b336f615e59ed18ec0

> #### **Performance:**
> Language such as "quick, easy, and free of bugs" is used in a few places, but I don't see any specifics on how the custom-made optimizers perform. I am familiar enough with many of them that I can attest to their veracity in fitting parameters (and they are well-covered by the package's `testthat`), but speed or efficiency haven't been documented. I do agree that `univariateML` provides a lot of convenience. But if language like "quick" is going to be used, it should be backed up with evidence.

I added an example with the `beta` distribution to the readme. Anyway, I've tried to remove all wording about the speed of the implementations, as that is not the focus of this package: The focus of the package is to lessen the time spent coding the implementations, making plots, etc.

(In my experience the usual way to do univariate maximum likelihood is to do an `nlm` or `optim` variant such as in the readme, maybe using the `mle` wrapper of `stats4`. The analytic solutions will always be faster than this and the custom made solvers in this package can be demonstrated to be faster on a case by case basis.)

Removed the reference to "free of bugs.".

> **Functionality documentation:**
> Please link to your documentation site more directly within the README of the github repository. I realize there's a Documentation header inside the readme, but I'd also link to the main site ([https://univariateml.netlify.com/index.html](https://univariateml.netlify.com/index.html)) at the top of the github repo (i.e. above the Topics) as well as within the Overview section.

Good idea, added link to the Github header and the overview section as requested.

> Within the documentation site, please re-order the Articles dropdown menu so that the vignettes are listed in an order that's friendly to new users (i.e. put the Overview vignette first). This can be done by further editing the  `_pkgdown.yml`  in the root directory.

Thanks for the tip! Done it and uploaded it to Netlify.

> Within the Distributions vignette, I'd make it a little more clear that the Estimator column refers to functions within your  `univariateML`  package. I think this can be done via renaming the column and reordering the column names to: Name, univariateML Function, Parameters, Support, Package.

Okay! Done it and uploaded it to Netlify.

> In the Copula vignette, the last line of the first section notes that the package can be used to "speed up task 1" (fitting marginal distributions), but the specifics of how speedy this is (e.g. with run times) or how it compares to approaches via other packages isn't shown. Again, I think that if speed or efficiency are invoked, there should be evidence shown. Or the language could just be revised.

Thanks; I just revised the language.

> **State of the field:**
> I'd recommend expanding this further. Right now, the paper mentions how this package compares to  `stats4::mle()`  but does not really go into enough detail about the specific benefits of using  `univariateML`'s functions.
>
>Within the paper, can you give at least one example of a custom-made optimizer within  `univariateML`  and show how this optimizer's performance fares against what is built into base R or another package? How have other functions let us down?

Thanks! I forgot to check this. Updated the paper accordingly. I've removed claims about speed from the paper.

> I suspect that the ability to fit each distribution is already available through other R packages (and this seems to be somewhat confirmed by the table in the Distributions vignette). If there are any distributions that are not covered by other packages, please make this explicit in the paper by specifying the name of the  `univariateML`  estimator function.

`nlm` and `optim` will work with any `d***` function pretty easily; so in a sense all of the estimators are available in the packages implementing the densities themselves.

> I also think there is merit to incorporating more of a "one stop shop" argument in here. It's convenient to have all of these functions within the same package, especially as they feed nicely into the "Data analysis aids" functions, which I'd argue are are the most important parts of the package. So I recommend that the language of the paper be further adjusted to tell the reader why it's good to use this package instead of relying on what is already out there.

I agree that's one of main benefits of the package, if not the main purpose. I have tried to make this clearer in the paper.

#### Additional comments:
> * I suggest inserting a `_` within the names of the maximum likelihood estimator functions. I had some trouble distinguishing function names like `mllnorm()` from `mlnorm()` as I tested things out. Having the `_` could help here, e.g. `ml_lnorm()` vs `ml_norm()` is easier for me to distinguish.

I agree it would be clearer! The main reason for using `ml***` instead of `ml_***` is consistency with the naming convention for densities (`d***`), distribution functions (`p***`), quantile functions (`q***`) and random deviate generation functions (`r***`). In addition, there are several packages with `e***` functions (expectation) and `v***` functions (variance).

> * Any interest in creating a delta AIC function? I.e. take the `data.frame` made by `stats::AIC()`, find the minimum AIC value and calculate the delta of each model's AIC from that? Not sure if this already exists elsewhere, but I assume not since you custom-wrote a few lines in your vignettes. But it could be a nice addition to the "Data analysis aids" set of functions that basically supersedes `stats::AIC()` by giving the AICs and also deltas (and maybe AICcs if the user prefers?)

This is probably a good idea but I would like to keep the package as lightweight as I can. Another idea in the same vein is to implement the whole "choose best AIC" model as in the copula vignette, but I feel this is a little beyond the core features of the package. (In a sense the package doesn't "support" AIC either, it just exports objects a `logLik` generic.)

> * I also have a suggestion that I'll leave entirely optional at this point: You might give some thought to the 'ecosystem' approach of integrating `univariateML` with your previous `kdensity` package. Can objects made from one package be used by the other in a convenient way? If so, it might be nice to have a 4th vignette demonstrating this and/or some language in the JOSS paper.

Yep, and well observed! The plan is to integrate `kdensity` with `univariateML` once `univariateML` is on CRAN. `kdensity` depends on the density functions `d***` and estimator functions such as `ml***`, and will work together with univariateML without any problems! (One can hope...)

> #### Some relatively minor spelling/grammatical/formatting things:
> * Within the Overview vignette, the first line does not resolve into a bulleted list (at least on the rendered Netlify site)
> * Also in the Overview, section "Comparing Many Models with AIC": please revise the line "The generic in `R` can take multiple models, and the lower the the better" so that it's clearer to your reader that the second half of the sentence refers specifically to AIC values.
> * Also in the Overview vignette, section "Confidence Intervals", sentence 2: Change "Do do" to "To do"
> * In the "Copula Modeling" vignette: "...the copula is know,..." should be "...the copula is known,..."
> * `spelling::spell_check_package()` yields mostly false positives, but there are a couple items (e.g. "similarily" in `ProbabilityPlots.Rd:68`) that merit correction

Thanks! These are fixed in several commits.
