### 2019-12-03 #1863 @vbaliga
Hi @JonasMoss and @arfon,

I took a look through all the changes and Jonas' responses -- very happy with what I see! I just checked off all the items on the review list to reflect that I'm happy with the current state and have no further feedback. I did leave my original review comments in; please let me know if they should be removed.

Congrats, Jonas ✨:tada: -- it's all nicely put together and like I said before, I can see myself using it.

Cheers,
Vikram
🐢
